data_science/requirements.txt
torch
torchvision
requests
pandas
BeautifulSoup4
PyPDF2
mysql-connector-python 
pytesseract